logging feature simplified
