--- src/main/java/com/cocoon/annotation/ExecutionTimeLog.java
@@ -1,11 +1,0 @@
-package com.cocoon.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ExecutionTimeLog {
-}
