--- src/main/java/com/cocoon/dto/LogSearchDTO.java
@@ -1,23 +1,0 @@
-package com.cocoon.dto;
-
-import com.cocoon.enums.ActionType;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.springframework.format.annotation.DateTimeFormat;
-
-import java.time.LocalDate;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-public class LogSearchDTO {
-    private UserDTO user;
-    private ActionType actionType;
-    @DateTimeFormat(pattern = "yyyy-MM-dd")
-    private LocalDate startDate;
-    @DateTimeFormat(pattern = "yyyy-MM-dd")
-    private LocalDate endDate;
-}
