--- src/main/java/com/cocoon/entity/UserLog.java
@@ -1,27 +1,0 @@
-package com.cocoon.entity;
-
-import com.cocoon.enums.ActionType;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@Table(name = "user_log")
-public class UserLog {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private User user;
-
-    private ActionType actionType;
-    private LocalDateTime dateTime = LocalDateTime.now();
-}
